Don't try to (re)load projects during `postProcessExecutionPlan`
