+++ subprojects/core/src/main/java/org/gradle/execution/TaskNameResolvingBuildTaskScheduler.java
@@ -16,6 +16,7 @@
 package org.gradle.execution;
 
 import org.gradle.TaskExecutionRequest;
+import org.gradle.api.NonNullApi;
 import org.gradle.api.internal.GradleInternal;
 import org.gradle.execution.commandline.CommandLineTaskParser;
 import org.gradle.execution.plan.ExecutionPlan;
@@ -43,7 +44,7 @@ public class TaskNameResolvingBuildTaskScheduler implements BuildTaskScheduler {
     @Override
     public void scheduleRequestedTasks(GradleInternal gradle, @Nullable EntryTaskSelector selector, ExecutionPlan plan) {
         if (selector != null) {
-            selector.applyTasksTo(new DefaultEntryTaskSelectorContext(gradle, taskSelector), plan);
+            selector.applyTasksTo(new EntryTaskSelectorContext(gradle), plan);
         }
         List<TaskExecutionRequest> taskParameters = gradle.getStartParameter().getTaskRequests();
         for (TaskExecutionRequest taskParameter : taskParameters) {
@@ -52,6 +53,25 @@ public class TaskNameResolvingBuildTaskScheduler implements BuildTaskScheduler {
                 LOGGER.info("Selected primary task '{}' from project {}", taskSelection.getTaskName(), taskSelection.getProjectPath());
                 plan.addEntryTasks(taskSelection.getTasks());
             }
+        }
+    }
+
+    @NonNullApi
+    private class EntryTaskSelectorContext implements EntryTaskSelector.Context {
+        final GradleInternal gradle;
+
+        public EntryTaskSelectorContext(GradleInternal gradle) {
+            this.gradle = gradle;
+        }
+
+        @Override
+        public TaskSelection getSelection(String taskPath) {
+            return taskSelector.resolveTaskName(taskPath);
+        }
+
+        @Override
+        public GradleInternal getGradle() {
+            return gradle;
         }
     }
 }
